Update ecfr-parser tests
