+++ solution/parser/lib/eregs/config_test.go
@@ -8,6 +8,59 @@ import (
 	"github.com/go-test/deep"
 )
 
+func TestGetLogLevel(t *testing.T) {
+	testTable := []struct {
+		Name     string
+		Input    string
+		Expected log.Level
+	}{
+		{
+			Name:     "test-warn",
+			Input:    "warn",
+			Expected: log.WarnLevel,
+		},
+		{
+			Name:     "test-fatal",
+			Input:    "fatal",
+			Expected: log.FatalLevel,
+		},
+		{
+			Name:     "test-error",
+			Input:    "error",
+			Expected: log.ErrorLevel,
+		},
+		{
+			Name:     "test-info",
+			Input:    "info",
+			Expected: log.InfoLevel,
+		},
+		{
+			Name:     "test-debug",
+			Input:    "debug",
+			Expected: log.DebugLevel,
+		},
+		{
+			Name:     "test-trace",
+			Input:    "trace",
+			Expected: log.TraceLevel,
+		},
+		{
+			Name:     "test-default",
+			Input:    "not a valid level",
+			Expected: log.WarnLevel,
+		},
+	}
+
+	for _, tc := range testTable {
+		Run(t, tc.Name, func(t *testing.T) {
+			out := getLogLevel(tc.Input)
+			if out != tc.Expected {
+				t.Errorf("expected (%+v), received (%+v)", tc.Expected, out)
+			}
+		})
+	}
+}
+
 func TestSubchapterArgString(t *testing.T) {
 	arg := SubchapterArg{"one", "two"}
 	expected := "one-two"
